Convert augfuncs to float storage. This introduces a very small change in total energies (3.0e-8 Ha for C60) while saving a large amount of memory when USPP are used. Needs more testing though.

+++ Headers/ION.h
@@ -199,7 +199,7 @@ public:
     double init_spin_z;
 
     // Augmentation charges associated with this ion
-    std::vector<double> augfunc;
+    std::vector<float> augfunc;
     std::vector<double> augfunc_xyz[3];
 
     // An index array which maps the q-functions onto the 3-d grid associated with each processor.
